crypto/requirements.txt

pycrypto==2.6.1
six==1.10
